Split client-independent level functions from levelmanager
